+++ hadoop-ozone/tools/src/main/java/org/apache/hadoop/ozone/shell/volume/DeleteVolumeHandler.java
@@ -20,7 +20,9 @@ package org.apache.hadoop.ozone.shell.volume;
 
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.fs.ozone.RootedOzoneFileSystem;
 import org.apache.hadoop.hdds.conf.OzoneConfiguration;
+import org.apache.hadoop.ozone.OFSPath;
 import org.apache.hadoop.ozone.client.OzoneBucket;
 import org.apache.hadoop.ozone.client.OzoneClient;
 import org.apache.hadoop.ozone.client.OzoneKey;
@@ -176,6 +178,11 @@ public class DeleteVolumeHandler extends VolumeHandler {
       OzoneConfiguration clientConf = new OzoneConfiguration(getConf());
       clientConf.set(FS_DEFAULT_NAME_KEY, hostPrefix);
       FileSystem fs = FileSystem.get(clientConf);
+      if (handleSpecialVolumeTmp(path, (RootedOzoneFileSystem) fs,
+          clientConf)) {
+        numberOfBucketsCleaned.getAndIncrement();
+        return true;
+      }
       if (!fs.delete(path, true)) {
         throw new IOException("Failed to delete bucket");
       }
@@ -186,6 +193,19 @@ public class DeleteVolumeHandler extends VolumeHandler {
       LOG.error("Could not clean bucket ", e);
       return false;
     }
+  }
+
+  private static boolean handleSpecialVolumeTmp(Path path,
+                                                RootedOzoneFileSystem fs,
+                                                OzoneConfiguration clientConf) {
+    String key = path.toUri().getPath().substring(1);
+    RootedOzoneFileSystem ofs = fs;
+    OFSPath ofsPath = new OFSPath(key, clientConf);
+    if (ofsPath.getBucketName().equals(OFSPath.md5Hex(ofs.getUsername())) ||
+        !ofsPath.isBucket()) {
+      return true;
+    }
+    return false;
   }
 
   private class BucketCleaner implements Runnable {
